+++ Go-Bee.novaextension/Scripts/config.js
@@ -10,6 +10,8 @@
 const keys = {
   lspFlavor: "gobee.lsp.flavor",
   lspPath: "gobee.lsp.path",
+  dapFlavor: "gobee.dap.flavor",
+  dapPath: "gobee.dap.path",
   goExec: "gobee.go.exec",
   goPath: "gobee.gopath",
   buildFlags: "gobee.buildFlags",
@@ -23,6 +25,8 @@ const keys = {
   version: "gobee.version", // our version so that other extensions can find us
   currentGoPls: "gobee.gopls.current",
   releaseGoPls: "gobee.gopls.release",
+  currentDelve: "gobee.delve.current",
+  releaseDelve: "gobee.delve.release",
 };
 
 module.exports = keys;