+++ Go-Bee.novaextension/Scripts/update.js
@@ -13,25 +13,35 @@ const Prefs = require("./prefs.js");
 const Paths = require("./paths.js");
 
 const extPath = nova.extension.globalStoragePath;
-let srvPath = nova.path.join(extPath, "gopls");
+
+const progs = {
+  gopls: {
+    name: "gopls",
+    repo: "golang/tools",
+    pkg: "golang.org/x/tools/gopls",
+    pfx: "gopls/",
+    rel: Config.releaseGoPls,
+    cur: Config.currentGoPls,
+    restart: true,
+  },
+  dlv: {
+    name: "dlv",
+    repo: "go-delve/delve",
+    pkg: "github.com/go-delve/delve/cmd/dlv",
+    pfx: "",
+    rel: Config.releaseDelve,
+    cur: Config.currentDelve,
+  },
+};
 
 async function checkForUpdate(force = false) {
-  let release = await GitHub.latest(force);
+  return Promise.all(
+    checkForUpdateProg(progs.dlv, force),
+    checkForUpdateProg(progs.gopls, force)
+  );
+}
 
-  let next = nova.config.get(Config.releaseGoPls);
-  if (release != null && "gopls/" + next != release.tag_name) {
-    next = release.tag_name;
-    if (next.startsWith("gopls/")) {
-      next = next.substr(6);
-    }
-
-    nova.config.set(Config.releaseGoPls, next);
-  }
-  let curr = nova.config.get(Config.currentGoPls);
-  if (curr) {
-    srvPath = nova.path.join(nova.extension.globalStoragePath, `gopls`);
-  }
-
+async function checkForGo() {
   let goexec = Prefs.getConfig(Config.goExec);
 
   if (goexec == null) {
@@ -47,31 +57,59 @@ async function checkForUpdate(force = false) {
     }
   }
 
-  if (!nova.fs.access(goexec, nova.fs.X_OK)) {
+  if (!goexec || !nova.fs.access(goexec, nova.fs.X_OK)) {
     Messages.showNotice(Catalog.msgNeedGoTitle, Catalog.msgNeedGoBody);
+    return null;
+  }
+  return goexec;
+}
+
+async function checkForUpdateProg(prog, force = false) {
+  let goexec = await checkForGo();
+  if (!goexec) {
     return false;
   }
+  let release = await GitHub.latest(prog.name, prog.repo, force);
 
-  if (curr && !nova.fs.access(srvPath, nova.fs.X_OK)) {
+  let next = nova.config.get(prog.rel);
+  if (release != null && prog.pfx + next != release.tag_name) {
+    next = release.tag_name;
+    if (prog.pfx.length > 0 && next.startsWith(prog.pfx)) {
+      next = next.substr(prog.pfx.length);
+    }
+
+    nova.config.set(prog.rel, next);
+  }
+  let curr = nova.config.get(prog.cur);
+  let exePath = "";
+  if (curr) {
+    exePath = nova.path.join(extPath, prog.name);
+  }
+
+  if (curr && !nova.fs.access(exePath, nova.fs.X_OK)) {
     console.error("We should have had a binary, but it appears to be missing.");
     curr = null;
   }
 
-  if (curr == next && nova.fs.access(srvPath, nova.fs.X_OK)) {
-    Messages.showNotice(Catalog.msgUpToDate, curr);
-    console.log("Language server is current.");
+  if (curr == next && nova.fs.access(exePath, nova.fs.X_OK)) {
+    Messages.showNotice(
+      Messages.getMsg(Catalog.msgUpToDate).replace("_PROG_", prog.name),
+      `${prog.name} ${curr}`
+    );
+    console.log(`Program ${prog.name} is current.`);
     return true;
   }
 
   let title = Messages.getMsg(
-    curr ? Catalog.msgNewLspTitle : Catalog.msgMissingLspTitle
+    curr ? Catalog.msgNewComponentTitle : Catalog.msgMissingComponentTitle
   );
   let text = Messages.getMsg(
-    curr ? Catalog.msgNewLspBody : Catalog.msgMissingLspBody
+    curr ? Catalog.msgNewComponentBody : Catalog.msgMissingComponentBody
   );
   let n = new NotificationRequest("autoUpdate");
   n.title = title;
   n.body = text
+    .replace("_PROG_", prog.name)
     .replace("_VERSION_", next)
     .replace("_OLD_VERSION_", curr ?? "none");
   n.actions = [
@@ -87,24 +125,26 @@ async function checkForUpdate(force = false) {
   }
 
   let options = {
-    args: ["install", "golang.org/x/tools/gopls@" + next],
-    env: { GOBIN: nova.extension.globalStoragePath },
+    args: ["install", prog.pkg + "@" + next],
+    env: { GOBIN: extPath },
   };
 
   let proc = new Process(goexec, options);
   proc.onDidExit((status) => {
     if (status == 0) {
       // notify watchers (should cause a reboot)
-      State.emitter.emit(State.events.onUpdate);
-      nova.config.set(Config.currentGoPls, next);
+      if (prog.restart) {
+        State.emitter.emit(State.events.onUpdate);
+      }
+      nova.config.set(prog.cur, next);
     } else {
-      Messages.showError(Catalog.msgLspUpdateFailed);
+      Messages.showError(Catalog.msgComponentUpdateFailed);
     }
   });
   try {
     proc.start();
   } catch (err) {
-    Messages.showError(Catalog.msgLspUpdateFailed);
+    Messages.showError(Catalog.msgComponentUpdateFailed);
   }
 }
 
@@ -113,8 +153,8 @@ async function checkForUpdate(force = false) {
 async function checkForUpdateCmd() {
   if (Prefs.getConfig(Config.lspFlavor) != "auto") {
     Messages.showNotice(
-      Catalog.msgLspIsNotAutoTitle,
-      Catalog.msgLspIsNotAutoBody
+      Catalog.msgComponentIsNotAutoTitle,
+      Catalog.msgComponentIsNotAutoBody
     );
     return;
   }
@@ -129,14 +169,13 @@ async function checkForUpdateSilent() {
   // If we should check for new versions at start up, try to download from
   // GitHub releases.  But we don't do this if the server is disabled or we
   // are using a custom server.
-  if (Prefs.getConfig(Config.lspFlavor) != "auto") {
-    return;
-  }
   if (Prefs.getConfig(Config.checkForUpdates)) {
-    // if it doesn't work, don't bother warning about it.
-    try {
-      await checkForUpdate();
-    } catch (error) {}
+    if (Prefs.getConfig(Config.lspFlavor) == "auto") {
+      // if it doesn't work, don't bother warning about it.
+      try {
+        await checkForUpdateProg(progs.gopls);
+      } catch (error) {}
+    }
   }
 }